# perf: improve string operations