--- a/lib/pretty-print.js
+++ b/lib/pretty-print.js
@@ -17,7 +17,7 @@ function printObjectTree (obj, parentPrefix = '') {
     const childPrefix = isLast ? '    ' : '│   '
 
     const nodeData = value[treeDataSymbol] || ''
-    const prefixedNodeData = nodeData.split('\n').join('\n' + parentPrefix + childPrefix)
+    const prefixedNodeData = nodeData.replaceAll('\n', '\n' + parentPrefix + childPrefix)
 
     tree += parentPrefix + nodePrefix + key + prefixedNodeData + '\n'
     tree += printObjectTree(value, parentPrefix + childPrefix)