--- a/lib/strategies/accept-version.js
+++ b/lib/strategies/accept-version.js
@@ -18,7 +18,7 @@ SemVerStore.prototype.set = function (version, store) {
   if (typeof version !== 'string') {
     throw new TypeError('Version should be a string')
   }
-  let [major, minor, patch] = version.split('.')
+  let [major, minor, patch] = version.split('.', 3)
 
   if (isNaN(major)) {
     throw new TypeError('Major version must be a numeric value')