@Fdawgs 2024-12-18
@gurgunday Is `.replaceAll()` slower than `.replace(//g)`?