@gurgunday 2024-12-18
I try to avoid RegExps in general, but the difference between them was nonsignificant when I last benched